#330 - refactor(frontend): Fix and add tests
