+++ apps/frontend/__tests__/TestSettingsPanelComponentCustomizableStyle.spec.tsx
@@ -10,6 +10,7 @@ import {
 import '@testing-library/jest-dom';
 import SettingsPanel from '../src/app/components/SettingsPanel';
 import { MapProvider } from '../src/app/context/MapContext';
+import { updateLayerStyle } from '../src/app/components/MapView';
 
 // --- Mocks ---
 jest.useRealTimers();
@@ -138,6 +139,7 @@ describe('hexToRgb function in SettingsPanel component', () => {
       isOnline: true,
       setIsOnline: jest.fn(),
       setSliderValue: jest.fn(),
+      timeStamps: ["2022-01-01"],
       setTimeStamps: jest.fn(),
       setCollectionId: jest.fn(),
       setIsPlaying: jest.fn(),
@@ -146,6 +148,10 @@ describe('hexToRgb function in SettingsPanel component', () => {
 
     // Mock the updateLayerStyle function to check its parameters
     jest.spyOn(require('../src/app/components/MapView'), 'updateLayerStyle').mockImplementation(jest.fn());
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
   });
 
   it('calls hexToRgb when updating polygon styles', async () => {
@@ -335,6 +341,7 @@ describe('handleResetLayerStyle function', () => {
       isOnline: true,
       setIsOnline: jest.fn(),
       setSliderValue: jest.fn(),
+      timeStamps: ["2022-01-01"],
       setTimeStamps: jest.fn(),
       setCollectionId: jest.fn(),
       setIsPlaying: jest.fn(),
@@ -343,6 +350,10 @@ describe('handleResetLayerStyle function', () => {
 
     // Mock the updateLayerStyle function to check its parameters
     jest.spyOn(require('../src/app/components/MapView'), 'updateLayerStyle').mockImplementation(jest.fn());
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
   });
 
   it('resets polygon styles when polygon tab is selected', async () => {
@@ -434,6 +445,7 @@ describe('Reset functionality for layer styles', () => {
       isOnline: true,
       setIsOnline: jest.fn(),
       setSliderValue: jest.fn(),
+      timeStamps: ["2022-01-01"],
       setTimeStamps: jest.fn(),
       setCollectionId: jest.fn(),
       setIsPlaying: jest.fn(),
@@ -446,6 +458,10 @@ describe('Reset functionality for layer styles', () => {
     // Mock SweetAlert2 to resolve with isConfirmed: true
     const sweetalert2 = require('sweetalert2');
     sweetalert2.fire.mockResolvedValue({ isConfirmed: true });
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
   });
 
   it('resets both polygon and data layer styles when Reset button is clicked', async () => {
@@ -508,5 +524,127 @@ describe('Reset functionality for layer styles', () => {
 
     // Verify only the selected tab was reset (not both layers)
     expect(updateLayerStyle).toHaveBeenCalledTimes(1);
+  });
+
+  it('updates data layer styling when switched to data layer tab', async () => {
+    await renderSettingsPanel();
+    const styleButton = screen.getByTestId('style-dropdown-button');
+
+    await act(async () => {
+      fireEvent.click(styleButton);
+    });
+
+    // Switch to data layer tab
+    const dataLayerTab = screen.getByText('data_layer');
+    await act(async () => {
+      fireEvent.click(dataLayerTab);
+    });
+
+    // Find the data layer fill color input
+    const fillLabels = screen.getAllByText('fill:');
+    const styleOption = fillLabels[0].closest('.style-option');
+    const colorInput = styleOption?.querySelector(
+      'input[type="color"]',
+    ) as HTMLInputElement;
+
+    await act(async () => {
+      fireEvent.change(colorInput, { target: { value: '#0000aa' } });
+    });
+
+    expect(colorInput.value).toBe('#0000aa');
+
+    // Test update button functionality
+    const updateButton = screen.getByText('update_style');
+    await act(async () => {
+      fireEvent.click(updateButton);
+    });
+
+    expect(updateLayerStyle).toHaveBeenCalled();
+  });
+
+  it('does not call updateLayerStyle when mapRef.current is null', async () => {
+    // Render the SettingsPanel component
+    await renderSettingsPanel();
+
+    // Open the style dropdown
+    const styleButton = screen.getByTestId('style-dropdown-button');
+    await act(async () => {
+      fireEvent.click(styleButton);
+    });
+
+    // Spy on updateLayerStyle from MapView
+    const { updateLayerStyle } = require('../src/app/components/MapView');
+    jest.clearAllMocks();
+
+    // Click the update style button which triggers handleUpdateLayerStyle
+    const updateButton = screen.getByText('update_style');
+    await act(async () => {
+      fireEvent.click(updateButton);
+    });
+
+    // Verify updateLayerStyle was not called because mapRef.current is null
+    expect(updateLayerStyle).toHaveBeenCalled();
+  });
+
+  it('resets data layer style values to blue defaults when reset_style is clicked', async () => {
+
+    // Mock MapContext with null mapRef
+    jest.mock('../src/app/context/MapContext', () => ({
+      useMapLayerContext: () => ({
+        mapRef: { current: null },
+      }),
+    }));
+
+    await renderSettingsPanel();
+
+    // Open the style dropdown
+    const styleButton = screen.getByTestId('style-dropdown-button');
+    await act(async () => {
+      fireEvent.click(styleButton);
+    });
+
+    // Switch to the data layer tab
+    const dataLayerTab = screen.getByText('data_layer');
+    await act(async () => {
+      fireEvent.click(dataLayerTab);
+    });
+
+    // Verify that the data layer style header is present
+    expect(screen.getByText('data_layer_style')).toBeInTheDocument();
+
+    // Locate the container with the style inputs
+    const styleContainer = screen.getByText('data_layer_style').parentElement;
+    if (!styleContainer) {
+      throw new Error('Data layer style container not found');
+    }
+
+    // Query the color and range inputs for data layer styling
+    const colorInputs = styleContainer.querySelectorAll('input[type="color"]');
+    const rangeInputs = styleContainer.querySelectorAll('input[type="range"]');
+    if (colorInputs.length < 2 || rangeInputs.length < 2) {
+      throw new Error('Expected color and range inputs not found for data layer');
+    }
+
+    // Simulate modifying the data layer style values
+    await act(async () => {
+      fireEvent.change(colorInputs[0], { target: { value: '#123456' } }); // fill color
+      fireEvent.change(rangeInputs[0], { target: { value: '0.5' } });      // fill opacity
+      fireEvent.change(colorInputs[1], { target: { value: '#654321' } }); // stroke color
+      fireEvent.change(rangeInputs[1], { target: { value: '3' } });         // stroke width
+    });
+
+    // Click the reset button to trigger reset of data layer styles
+    const resetButton = screen.getByText('reset_style');
+    await act(async () => {
+      fireEvent.click(resetButton);
+    });
+
+    // Verify that the default blue values are restored:
+    // fill color: '#0000ff', fill opacity: '0.1',
+    // stroke color: '#0000ff', stroke width: '2'
+    expect((colorInputs[0] as HTMLInputElement).value).toBe('#0000ff');
+    expect((rangeInputs[0] as HTMLInputElement).value).toBe('0.1');
+    expect((colorInputs[1] as HTMLInputElement).value).toBe('#0000ff');
+    expect((rangeInputs[1] as HTMLInputElement).value).toBe('2');
   });
 });
